pinctrl: msm: Remove explicit barriers from mmio ops where unneeded
For the vast majority of mmio operations in this driver, explicit memory
barriers aren't needed either because a data dependency between a read
and write already exists, or because of the presence of the spin locks
which execute a full memory barrier.

Removing all the unneeded explicit barriers considerably reduces
overhead for pinctrl operations, which in turn benefits things like i2c.

Change-Id: I4576df1dc62f75dc98a16bd9b49fde0c1664ecbf
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Yahya Wessam <yahyawessam2002@gmail.com>